+++ src/main/java/com/getindata/connectors/http/internal/table/lookup/HttpLookupTableSource.java
@@ -2,6 +2,7 @@ package com.getindata.connectors.http.internal.table.lookup;
 
 import java.util.ArrayList;
 import java.util.List;
+import javax.annotation.Nullable;
 
 import lombok.extern.slf4j.Slf4j;
 import org.apache.flink.api.common.serialization.DeserializationSchema;
@@ -10,17 +11,19 @@ import org.apache.flink.table.api.DataTypes;
 import org.apache.flink.table.api.DataTypes.Field;
 import org.apache.flink.table.connector.Projection;
 import org.apache.flink.table.connector.format.DecodingFormat;
-import org.apache.flink.table.connector.source.AsyncTableFunctionProvider;
 import org.apache.flink.table.connector.source.DynamicTableSource;
 import org.apache.flink.table.connector.source.LookupTableSource;
 import org.apache.flink.table.connector.source.abilities.SupportsLimitPushDown;
 import org.apache.flink.table.connector.source.abilities.SupportsProjectionPushDown;
+import org.apache.flink.table.connector.source.lookup.AsyncLookupFunctionProvider ;
 import org.apache.flink.table.connector.source.lookup.LookupFunctionProvider;
+import org.apache.flink.table.connector.source.lookup.PartialCachingAsyncLookupProvider;
 import org.apache.flink.table.connector.source.lookup.PartialCachingLookupProvider;
 import org.apache.flink.table.connector.source.lookup.cache.LookupCache;
 import org.apache.flink.table.data.RowData;
 import org.apache.flink.table.factories.DynamicTableFactory;
 import org.apache.flink.table.factories.FactoryUtil;
+import org.apache.flink.table.functions.AsyncLookupFunction;
 import org.apache.flink.table.types.DataType;
 import org.apache.flink.table.types.logical.LogicalType;
 import org.apache.flink.table.types.logical.RowType;
@@ -48,6 +51,7 @@ public class HttpLookupTableSource
     private final DynamicTableFactory.Context dynamicTableFactoryContext;
 
     private final DecodingFormat<DeserializationSchema<RowData>> decodingFormat;
+    @Nullable
     private final LookupCache cache;
 
     public HttpLookupTableSource(
@@ -55,8 +59,7 @@ public class HttpLookupTableSource
             HttpLookupConfig lookupConfig,
             DecodingFormat<DeserializationSchema<RowData>> decodingFormat,
             DynamicTableFactory.Context dynamicTablecontext,
-            LookupCache cache) {
-
+            @Nullable LookupCache cache) {
         this.physicalRowDataType = physicalRowDataType;
         this.lookupConfig = lookupConfig;
         this.decodingFormat = decodingFormat;
@@ -100,39 +103,36 @@ public class HttpLookupTableSource
         PollingClientFactory<RowData> pollingClientFactory =
             createPollingClientFactory(lookupQueryCreator, lookupConfig);
 
-        // In line with the JDBC implementation and current requirements, we are only
-        // supporting Partial Caching for synchronous operations.
         return getLookupRuntimeProvider(lookupRow, responseSchemaDecoder, pollingClientFactory);
     }
 
     protected LookupRuntimeProvider getLookupRuntimeProvider(LookupRow lookupRow,
         DeserializationSchema<RowData> responseSchemaDecoder,
         PollingClientFactory<RowData> pollingClientFactory) {
+
+        HttpTableLookupFunction dataLookupFunction =
+                new HttpTableLookupFunction(
+                        pollingClientFactory,
+                        responseSchemaDecoder,
+                        lookupRow,
+                        lookupConfig
+                );
         if (lookupConfig.isUseAsync()) {
-            HttpTableLookupFunction dataLookupFunction =
-                    new HttpTableLookupFunction(
-                            pollingClientFactory,
-                            responseSchemaDecoder,
-                            lookupRow,
-                            lookupConfig
-                    );
-            log.info("Using Async version of HttpLookupTable.");
-            return AsyncTableFunctionProvider.of(
-                new AsyncHttpTableLookupFunction(dataLookupFunction));
-        } else {
-            CachingHttpTableLookupFunction dataLookupFunction =
-                    new CachingHttpTableLookupFunction(
-                            pollingClientFactory,
-                            responseSchemaDecoder,
-                            lookupRow,
-                            lookupConfig,
-                            cache
-                    );
+            AsyncLookupFunction asyncLookupFunction =
+                    new AsyncHttpTableLookupFunction(dataLookupFunction);
             if (cache != null) {
-                log.debug("PartialCachingLookupProvider; cache = " + cache);
+                log.info("Using async version of HttpLookupTable with cache.");
+                return PartialCachingAsyncLookupProvider.of(asyncLookupFunction, cache);
+            } else {
+                log.info("Using async version of HttpLookupTable without cache.");
+                return AsyncLookupFunctionProvider.of(asyncLookupFunction);
+            }
+        } else {
+            if (cache != null) {
+                log.info("Using blocking version of HttpLookupTable with cache.");
                 return PartialCachingLookupProvider.of(dataLookupFunction, cache);
             } else {
-                log.debug("Using LookupFunctionProvider.");
+                log.info("Using blocking version of HttpLookupTable without cache.");
                 return LookupFunctionProvider.of(dataLookupFunction);
             }
         }